+++ ContentPatcher/docs/release-notes.md
@@ -9,6 +9,9 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 
 -->
+## Upcoming release
+* Fixed location tokens unready during part of the save loading process.
+
 ## 1.24.5
 Released 30 November 2021 for SMAPI 3.13.0 or later.